# document/chatbi-openapi.yaml
openapi: 3.1.0
info:
  title: ChatBI Backend API
  version: 1.0.0
  description: |
    Backend contract consumed by the Dify ChatBI workflow.

    The API is fail-closed. LLM-generated SQL, physical table names, credentials,
    and permission expressions are never accepted as trusted execution input.
    The current Dify workflow sends `compiled_query` to the execute endpoint for
    compatibility; that field is deprecated and must be ignored as an authority.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /api/chatbi/context/load:
    post:
      tags: [Context]
      operationId: loadChatContext
      summary: Load trusted identity, workspace policy, and prior query context
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContextLoadRequest'
            examples:
              demo:
                value:
                  workspace_id: demo
                  conversation_id: conv_001
                  identity_token: demo-server-issued-token
      responses:
        '200':
          description: Context loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextLoadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/chatbi/dsl/validate:
    post:
      tags: [DSL]
      operationId: validateQueryDsl
      summary: Validate and normalize Query DSL 1.0
      description: |
        Validates protocol version, metric version and state, dimensions, filters,
        model compatibility, time range, query limits, workspace, and policy context.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DslValidateRequest'
      responses:
        '200':
          description: Validation decision; business invalidity is represented in the body status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DslValidateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chatbi/query/compile:
    post:
      tags: [Query]
      operationId: compileQuery
      summary: Compile validated semantic DSL and apply SQL safety controls
      description: |
        Generates SQL deterministically from published metric definitions and model
        relationships. SQL text is stored server-side and is not returned by default.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompileRequest'
      responses:
        '200':
          description: Compile and execution-gate decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chatbi/query/execute:
    post:
      tags: [Query]
      operationId: executeQuery
      summary: Execute a server-side compiled and authorized query
      description: |
        The server revalidates workspace, operator, metric versions, policy version,
        SQL fingerprint, expiration, and approval proof. It must never execute SQL
        supplied in the request body.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteRequest'
      responses:
        '200':
          description: Query result or terminal query status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Idempotency key conflicts with different query content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
        '504':
          description: Warehouse execution timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chatbi/result/profile:
    post:
      tags: [Result]
      operationId: profileResult
      summary: Build deterministic statistics, chart specification, and evidence
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
      responses:
        '200':
          description: Result profile and evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chatbi/interpretation/generate:
    post:
      tags: [Interpretation]
      operationId: generateInterpretation
      summary: Generate deterministic evidence-bound interpretation fallback
      description: |
        Builds a conservative Interpretation from server-generated Evidence. This
        endpoint is intended as a stable fallback when external LLM providers are
        unavailable or streaming responses fail.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterpretationGenerateRequest'
      responses:
        '200':
          description: Deterministic interpretation candidate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterpretationGenerateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Query, DSL, or profile does not match stored evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chatbi/reflection/validate:
    post:
      tags: [Reflection]
      operationId: validateInterpretation
      summary: Validate interpretation against deterministic evidence
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XTraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReflectionRequest'
      responses:
        '200':
          description: Reflection decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReflectionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
  parameters:
    XRequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Client-generated request identifier. The server generates one when absent.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    XTraceId:
      name: X-Trace-ID
      in: header
      required: false
      description: Distributed trace identifier propagated through Dify and backend services.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Must equal the query_id for the current Dify workflow.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    Unauthorized:
      description: Missing or invalid service token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller cannot access the requested workspace or resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request envelope or field validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal service failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/TraceContext'
        - type: object
          additionalProperties: false
          required: [status, code, message, retryable]
          properties:
            status:
              type: string
              const: ERROR
            code:
              type: string
              pattern: '^[A-Z][A-Z0-9_]{2,100}$'
            message:
              type: string
              maxLength: 1000
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
    ContextLoadRequest:
      type: object
      additionalProperties: false
      required: [workspace_id, conversation_id, identity_token]
      properties:
        workspace_id:
          type: string
          minLength: 1
          maxLength: 128
        conversation_id:
          type: string
          minLength: 1
          maxLength: 128
        identity_token:
          type: string
          minLength: 1
          maxLength: 4096
...
    ContextLoadResponse:
      allOf:
        - $ref: '#/components/schemas/TraceContext'
        - type: object
          additionalProperties: false
          required:
            - operator_id
            - allowed_domains
            - role_ids
            - row_policy_token
            - last_query_context
          properties:
            operator_id:
              type: string
            allowed_domains:
              type: array
              uniqueItems: true
              items:
                type: string
                enum: [sales, advertising]
            role_ids:
              type: array
              uniqueItems: true
              items:
                type: string
            row_policy_token:
              type: string
              description: Opaque short-lived policy reference; no policy expression is exposed.
            last_query_context:
              $ref: '#/components/schemas/LastQueryContext'
...
    QueryDsl:
      $ref: './query-dsl-v1.schema.json'
    DslValidateRequest:
      type: object
      additionalProperties: false
      required: [workspace_id, operator_id, row_policy_context, dsl]
      properties:
        workspace_id:
          type: string
        operator_id:
          type: string
        row_policy_context:
          type: object
          additionalProperties: true
        dsl:
          $ref: '#/components/schemas/QueryDsl'
...
    DslValidateResponse:
      allOf:
        - $ref: '#/components/schemas/TraceContext'
        - type: object
          additionalProperties: false
          required: [status, normalized_dsl, issues, message]
          properties:
            status:
              type: string
              enum: [VALID, CLARIFY, DENY, INVALID]
            normalized_dsl:
              oneOf:
                - $ref: '#/components/schemas/QueryDsl'
                - type: 'null'
            issues:
              type: array
              items:
                $ref: '#/components/schemas/ValidationIssue'
            message:
              type: string
              maxLength: 1000
    CompileRequest:
      type: object
      additionalProperties: false
      required: [workspace_id, operator_id, dsl, permission_context]
      properties:
        workspace_id:
          type: string
        operator_id:
          type: string
        dsl:
          $ref: '#/components/schemas/QueryDsl'
        permission_context:
          type: object
          additionalProperties: true
...
    CompileResponse:
      allOf:
        - $ref: '#/components/schemas/TraceContext'
        - type: object
          additionalProperties: false
          required:
            - status
            - query_id
            - sql_fingerprint
            - dsl_hash
            - metric_versions
            - lineage
            - estimated_cost
            - expires_at
            - message
          properties:
            status:
              type: string
              enum: [READY, NEED_APPROVAL, BLOCKED]
            query_id:
              type: string
            sql_fingerprint:
              type: string
              pattern: '^sha256:[a-fA-F0-9]{64}$'
            dsl_hash:
              type: string
              pattern: '^sha256:[a-fA-F0-9]{64}$'
            metric_versions:
              $ref: '#/components/schemas/MetricVersionMap'
            lineage:
              $ref: '#/components/schemas/Lineage'
            estimated_cost:
              $ref: '#/components/schemas/EstimatedCost'
            execution_token:
              type: string
              description: Short-lived signed token for READY queries. It contains no SQL text.
            expires_at:
              type: string
              format: date-time
            message:
              type: string
    ExecuteRequest:
      type: object
      additionalProperties: false
      required: [workspace_id, operator_id, query_id]
      properties:
        workspace_id:
          type: string
        operator_id:
          type: string
        query_id:
          type: string
        execution_token:
          type: string
          description: Preferred signed execution proof returned by the compiler.
        approval_token:
          type: string
          description: Required only for NEED_APPROVAL queries after approval.
        compiled_query:
          type: object
          additionalProperties: true
          deprecated: true
          description: |
            Compatibility field sent by the current Dify DSL. The server must ignore
            it as executable content and load the compiled query by query_id.
...
    ExecuteResponse:
      allOf:
        - $ref: '#/components/schemas/TraceContext'
        - type: object
          additionalProperties: false
          required:
            - query_id
            - status
            - columns
            - rows
            - row_count
            - execution_ms
            - cached
            - truncated
            - result_ref
            - data_quality
          properties:
            query_id:
              type: string
            status:
              type: string
              enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED, TIMED_OUT]
            columns:
              type: array
              maxItems: 100
              items:
                $ref: '#/components/schemas/ResultColumn'
            rows:
              type: array
              maxItems: 5000
              items:
                type: object
                additionalProperties: true
            row_count:
              type: integer
              minimum: 0
            execution_ms:
              type: integer
              minimum: 0
            cached:
              type: boolean
            truncated:
              type: boolean
            result_ref:
              oneOf:
                - type: string
                - type: 'null'
            data_quality:
              $ref: '#/components/schemas/DataQualitySummary'
            error:
              oneOf:
                - $ref: '#/components/schemas/ErrorResponse'
                - type: 'null'
    ProfileRequest:
      type: object
      additionalProperties: false
      required: [workspace_id, query_id, execution_result, dsl]
      properties:
        workspace_id:
          type: string
        query_id:
          type: string
        execution_result:
          $ref: '#/components/schemas/ExecuteResponse'
        dsl:
          $ref: '#/components/schemas/QueryDsl'
...